requirements.txt
# CommPlexCore/requirements.txt
# Domain: The Brain -- Gemini Flash, Vertex AI, ML dependencies
# Install: pip install -r requirements.txt --break-system-packages

# ── Vertex AI / Gemini Flash ──────────────────────────────
google-cloud-aiplatform>=1.49.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0

# ── GCP Storage / BigQuery (for standup archiving) ────────
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.17.0

# ── HTTP (for KillSwitch ntfy.sh alert) ──────────────────
requests>=2.31.0

# ── Config / Environment ──────────────────────────────────
python-dotenv>=1.0.0

# ── Type hints (Python 3.8 compat) ───────────────────────
typing-extensions>=4.9.0

# ── Optional: Tenacity for retry/backoff ──────────────────
tenacity>=8.2.0
